Refactor

+++ contracts/interface/IERC20.sol
@@ -2,17 +2,40 @@
 pragma solidity >=0.7.0;
 
 interface IERC20 {
-    event Approval(address,address,uint);
-    event Transfer(address,address,uint);
+    event Approval(address, address, uint256);
+    event Transfer(address, address, uint256);
+
     function name() external view returns (string memory);
+
     function decimals() external view returns (uint8);
-    function transferFrom(address,address,uint) external returns (bool);
-    function allowance(address,address) external view returns (uint);
-    function approve(address,uint) external returns (bool);
-    function transfer(address,uint) external returns (bool);
-    function balanceOf(address) external view returns (uint);
-    function nonces(address) external view returns (uint);  // Only tokens that support permit
-    function permit(address,address,uint256,uint256,uint8,bytes32,bytes32) external;  // Only tokens that support permit
-    function swap(address,uint256) external;  // Only Avalanche bridge tokens 
-    function swapSupply(address) external view returns (uint);  // Only Avalanche bridge tokens 
+
+    function transferFrom(
+        address,
+        address,
+        uint256
+    ) external returns (bool);
+
+    function allowance(address, address) external view returns (uint256);
+
+    function approve(address, uint256) external returns (bool);
+
+    function transfer(address, uint256) external returns (bool);
+
+    function balanceOf(address) external view returns (uint256);
+
+    function nonces(address) external view returns (uint256); // Only tokens that support permit
+
+    function permit(
+        address,
+        address,
+        uint256,
+        uint256,
+        uint8,
+        bytes32,
+        bytes32
+    ) external; // Only tokens that support permit
+
+    function swap(address, uint256) external; // Only Avalanche bridge tokens
+
+    function swapSupply(address) external view returns (uint256); // Only Avalanche bridge tokens
 }
